Move notify tests to its own file
